Part 1: FunctionOnManifold
