+++ include/roboptim/core/manifold-map/decorator/descriptive-wrapper.hh
@@ -51,7 +51,7 @@ namespace roboptim
     ///
     /// \param f input function.
     /// \param m the manifold describing the function's input vector.
-    DescriptiveWrapper (const U* f, const mnf::Manifold& m);
+    DescriptiveWrapper (U* f, const mnf::Manifold& m);
 
     /// \brief binds the function to the manifold
     ///
@@ -74,19 +74,19 @@ namespace roboptim
     }
 
     /// \brief const wrapped function getter
-    const U& fct ()
+    U& fct ()
     {
       return *fct_;
     }
 
-    static DescriptiveWrapper<U, V>* makeUNCHECKEDDescriptiveWrapper(const U* fct, const mnf::Manifold& manifold);
+    static DescriptiveWrapper<U, V>* makeUNCHECKEDDescriptiveWrapper(U* fct, const mnf::Manifold& manifold);
 
   private:
     /// \brief dimension check between the function and the descriptive manifold
     void checkDimension();
 
     ///\brief the function
-    const U*              fct_;
+    U*                    fct_;
     ///\brief the manifold
     const mnf::Manifold*  manifold_;
   };
